FilterBox clean up
